project2.0

+++ src/lab4_ianlagos_felipelin/Ravenclaw.java
@@ -2,7 +2,7 @@ package lab4_ianlagos_felipelin;
 
 public class Ravenclaw extends Equipo {
 
-    public Ravenclaw(int partidosG, int partidosP, int promedioA, int promedioV, int promedioF, String capitan) {
+    public Ravenclaw(int partidosG, int partidosP, double promedioA, double promedioV, double promedioF, String capitan) {
         super(partidosG, partidosP, promedioA, promedioV, promedioF, capitan);
     }
