+++ src/lab4_ianlagos_felipelin/Slytherin.java
@@ -2,7 +2,7 @@ package lab4_ianlagos_felipelin;
 
 public class Slytherin extends Equipo {
 
-    public Slytherin(int partidosG, int partidosP, int promedioA, int promedioV, int promedioF, String capitan) {
+    public Slytherin(int partidosG, int partidosP, double promedioA, double promedioV, double promedioF, String capitan) {
         super(partidosG, partidosP, promedioA, promedioV, promedioF, capitan);
     }
